CODE/INTEXapi/requirements.txt
argon2-cffi==19.2.0
cffi==1.14.0
Django==1.11.29
django-cors-headers==3.0.2
djangorestframework==3.9.4
enum34==1.1.10
psycopg2==2.8.5
pycparser==2.20
pytz==2019.3
six==1.14.0
